AWS 25.0 crate (#1658)
This is a proposal for another way to integrate the Ada Web Server as a crate.
The issue with the current approach is the Makefile-based environment used
by AWS and the fact that we have to completely build and then install AWS
to use it.

In this proposal, only the 'make setup' phase is done to let the AWS Makefile
do the important setup and prepare the required files in target directory.
That target directory is named by AWS by using `gcc -dumpmachine` which brings
other issues and to solve and fix things, that directory is moved to a fixed
directory `build` so that every target will use that fix value.

The installation and build is not made and instead we reference the two
important GNAT projects 'aws.gpr' (top-level) and `build/projects/aws_config.gpr`
which was generated by `make setup`.  The build is then handled by Alire
since it knowns the AWS projects.  The AWS GNAT projects rely on the following
important variables which are set through the makefiles and we can set
them for each target as follows:

TGT_DIR = "../build"
TARGET = "build"

It is a proposal in the hope it will solve the AWS crate integration issues....